Finished refactoring and removed circurlar dependencies
