+++ cr_module/__init__.py
@@ -5,12 +5,5 @@ __author__ = "Ricardo Bartels <ricardo.bartels@telekom.de>"
 __description__ = "Check Redfish Plugin"
 __license__ = "MIT"
 
-__long_description__ = """
-This is a monitoring/inventory plugin to check components and
-health status of systems which support Redfish.
-It will also create a inventory of all components of a system.
-
-R.I.P. IPMI
-"""
 
 # EOF
